move base_cli to lib/inspec
It is not a disconnected library, but a core component of inspec. Fix its location.
