+++ lib/inspec.rb
@@ -19,7 +19,7 @@ require 'inspec/runner'
 require 'inspec/shell'
 
 # all utils that may be required by plugins
-require 'utils/base_cli'
+require 'inspec/base_cli'
 require 'inspec/fetcher'
 require 'inspec/source_reader'
 require 'inspec/resource'
