+++ lib/inspec/base_cli.rb
